refactor(release): move digestForPlan to release-contract, free the executor of proof
`digestForPlan` is a pure plan content-identity hash with no proof-specific
logic, yet it lived in proof.ts — forcing the executor, artifact, and
reconciler layers to import the proof module solely to compute a plan
digest, contradicting the documented "executor stays proof-blind" contract
(execute.ts:120-128).

Move `digestForPlan` + its private `PlanForDigest` struct to
release-contract.ts, colocated with `PlanDigest`/`digestPlanBody`, which
already owns plan-identity digesting and depends on digest.ts. The added
`Plan` import is type-only, so the existing plan.ts→release-contract.ts
value edge introduces no runtime cycle.

- execute.ts now has zero proof-module imports (proof-blind is literal).
- artifact.ts/reconciler.ts re-point to release-contract.js.
- The four `Api.Proof.digestForPlan` call sites (promise-api, apply, prune,
  archive) move to `Api.ReleaseContract.digestForPlan` — hard cut, no
  back-compat re-export.
- Digest output is byte-identical (same body, same short-circuit).

Task #6 of the @kitz/release SoC burn-down.

Session-Id: 4fb7ce2b-d9cc-4f7e-83f2-901e1d1fad5e
