+++ packages/release/src/api/executor/_.test.ts
@@ -7,7 +7,7 @@ import { describe, expect, test } from 'bun:test'
 import { Test } from '@kitz/test'
 import { Effect, FileSystem, Ref } from 'effect'
 import { execute, executeObservable, toPayload } from './execute.js'
-import { digestForPlan } from '../proof.js'
+import { digestForPlan } from '../release-contract.js'
 import { PublishIntent, publishIntentFromSemantics } from '../release-contract.js'
 import { resolvePublishSemantics } from '../publishing.js'
 import { Plan } from '../planner/models/plan.js'
